Manipulated testing to accomodate changes
