+++ test/test_command.py
@@ -1,0 +1,103 @@
+import sys
+import unittest
+
+from image_titler import command
+
+
+class TestCommand(unittest.TestCase):
+    pass
+
+
+class TestParseInput(TestCommand):
+
+    def setUp(self) -> None:
+        sys.argv = sys.argv[:1]  # clears args for each test
+
+    def test_default(self):
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, None)
+
+    def test_title(self):
+        sys.argv.append("-t")
+        sys.argv.append("Hello World")
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, "Hello World")
+
+    def test_path(self):
+        sys.argv.append("-p")
+        sys.argv.append("path/to/stuff")
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, "path/to/stuff")
+        self.assertEqual(args.tier, "")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, None)
+
+    def test_output_path(self):
+        sys.argv.append("-o")
+        sys.argv.append("path/to/stuff")
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "")
+        self.assertEqual(args.output_path, "path/to/stuff")
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, None)
+
+    def test_logo_path(self):
+        sys.argv.append("-l")
+        sys.argv.append("path/to/stuff")
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, "path/to/stuff")
+        self.assertEqual(args.title, None)
+
+    def test_batch(self):
+        sys.argv.append("-b")
+        args = command.parse_input()
+        self.assertEqual(args.batch, True)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, None)
+
+    def test_tier_premium(self):
+        sys.argv.append("-r")
+        sys.argv.append("premium")
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "premium")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, None)
+
+    def test_tier_free(self):
+        sys.argv.append("-r")
+        sys.argv.append("free")
+        args = command.parse_input()
+        self.assertEqual(args.batch, False)
+        self.assertEqual(args.path, None)
+        self.assertEqual(args.tier, "free")
+        self.assertEqual(args.output_path, None)
+        self.assertEqual(args.logo_path, None)
+        self.assertEqual(args.title, None)
+
+
+if __name__ == '__main__':
+    unittest.main()
